Added prettier with all changes that are needed
This addes prettier support with mostly default config values in
prettier. The changes to the default are "singleQuote: true",
"semi: true" and "bracketSpacing: true" these changes where needed to
still have the same formating as it currently is.

Eslint changes are: That the extends as 3 new entrys "prettier",
"prettier/@typescript-eslint" and "prettier/vue" these are needed so
that the formating from prettier does not throw erros when the source
code is linted.